requirements.txt
scikit-learn==1.2.2
streamlit_option_menu
scikit-learn
twilio
python-dotenv